Update type locations
