+++ src/client/sequencerClient.ts
@@ -7,9 +7,10 @@ import {
   GetTransactionReceiptResponse,
   Hash,
   SendRawTransactionResponse,
+  SequencerClient,
+  SequencerClientConfig,
   TransactionStatus,
 } from "@/types";
-import { createNodeClient, NodeClient } from "./nodeClient";
 import { Client, HTTPTransport, RequestManager } from "@open-rpc/client-js";
 import {
   formatBlockTagOrNumber,
@@ -18,30 +19,7 @@ import {
   formatGasPriceResponse,
 } from "@/types/formatters";
 import { DEFAULTS } from "@/constants";
-
-export interface SequencerClient extends NodeClient {
-  sendRawTransaction: ({ data }: { data: Data }) => Promise<SendRawTransactionResponse>;
-
-  waitForTransactionInclusion: ({ hash }: { hash: Hash }) => Promise<GetTransactionReceiptResponse>;
-
-  gasPrice: () => Promise<GasPriceResponse>;
-
-  estimateGas: ({ txData }: { txData: Data }) => Promise<EstimateGasResponse>;
-
-  estimateL1DataFee: ({
-    txData,
-    block,
-  }: {
-    txData: Data;
-    block?: BlockTagOrNumber;
-  }) => Promise<EstimateL1DataFeeResponse>;
-}
-
-export interface SequencerClientConfig {
-  url: string;
-  pollingIntervalMs?: number;
-  pollingRetries?: number;
-}
+import { createNodeClient } from "./nodeClient";
 
 export function createSequencerClient(config: SequencerClientConfig): SequencerClient {
   const {
