+++ src/index.ts
@@ -1,5 +1,4 @@
 export * from "./account";
 export * from "./client";
-export * from "./constants";
 export * from "./types";
 export * from "./transaction";
