+++ src/types/index.ts
@@ -1,4 +1,5 @@
 export * from "./block";
+export * from "./client";
 export * from "./input";
 export * from "./primitives";
 export * from "./response";
